rtpproxy:openembedded (Yocto) recipe
rtpproxy recipe that configures and builds rtpproxy recipe for
openembedded. This recipe has been tested on sumo branch and on PPC spe
architecture.
With the patch added, I have also fixed the broken-separation
autotools. If someone does not need the patch, it can remove it and just
use inherit autotools-brokensep

Signed-off-by: Panagiotis Tamtamis <panagiotis.tamtamis@atos.net>